# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lite-Harness API",
    "version": "0.1.0-alpha.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3210"
    }
  ],
  "components": {
    "securitySchemes": {
      "appToken": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "CreateRun": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "agent": {
            "minLength": 1,
            "maxLength": 128,
            "type": "string"
          },
          "workspace": {
            "minLength": 1,
            "maxLength": 128,
            "type": "string"
          },
          "session": {
            "minLength": 1,
            "maxLength": 128,
            "type": "string"
          },
          "input": {
            "minLength": 1,
            "maxLength": 1000000,
            "type": "string"
          },
          "budget": {
            "additionalProperties": false,
            "type": "object",
            "properties": {
              "maxTurns": {
                "minimum": 1,
                "maximum": 128,
                "type": "integer"
              },
              "maxToolCalls": {
                "minimum": 0,
                "maximum": 10000,
                "type": "integer"
              },
              "maxInputTokens": {
                "minimum": 1,
                "type": "integer"
              },
              "maxOutputTokens": {
                "minimum": 1,
                "type": "integer"
              },
              "maxCostUsd": {
                "minimum": 0,
                "type": "number"
              },
              "totalTimeoutMs": {
                "minimum": 100,
                "maximum": 86400000,
                "type": "integer"
              },
              "modelIdleTimeoutMs": {
                "minimum": 100,
                "maximum": 3600000,
                "type": "integer"
              },
              "commandTimeoutMs": {
                "minimum": 100,
                "maximum": 3600000,
                "type": "integer"
              }
            }
          }
        },
        "required": [
          "agent",
          "workspace",
          "input"
        ]
      },
...
      "CreateAgent": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "id": {
            "minLength": 1,
            "maxLength": 128,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "maxLength": 128,
            "type": "string"
          },
          "instructions": {
            "maxLength": 1000000,
            "type": "string"
          },
          "modelCapabilities": {
            "maxItems": 256,
            "type": "array",
            "items": {
              "minLength": 1,
              "maxLength": 128,
              "type": "string"
            }
          },
          "allowedTools": {
            "maxItems": 1000,
            "type": "array",
            "items": {
              "minLength": 1,
              "maxLength": 256,
              "type": "string"
            }
          },
          "defaultBudget": {
            "additionalProperties": false,
            "type": "object",
            "properties": {
              "maxTurns": {
                "minimum": 1,
                "maximum": 128,
                "type": "integer"
              },
              "maxToolCalls": {
                "minimum": 0,
                "maximum": 10000,
                "type": "integer"
              },
              "maxInputTokens": {
                "minimum": 1,
                "type": "integer"
              },
              "maxOutputTokens": {
                "minimum": 1,
                "type": "integer"
              },
              "maxCostUsd": {
                "minimum": 0,
                "type": "number"
              },
              "totalTimeoutMs": {
                "minimum": 100,
                "maximum": 86400000,
                "type": "integer"
              },
              "modelIdleTimeoutMs": {
                "minimum": 100,
                "maximum": 3600000,
                "type": "integer"
              },
              "commandTimeoutMs": {
                "minimum": 100,
                "maximum": 3600000,
                "type": "integer"
              }
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateWorkspace": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "id": {
            "minLength": 1,
            "maxLength": 128,
            "type": "string"
          },
          "mode": {
            "const": "managed",
            "type": "string"
          }
        }
      },
      "InboundEnvelope": {
        "type": "object",
        "required": [
          "deliveryId",
          "senderExternalId",
          "text"
        ],
        "properties": {
          "deliveryId": {
            "type": "string",
            "maxLength": 512
          },
          "senderExternalId": {
            "type": "string",
            "maxLength": 512
          },
          "conversationExternalId": {
            "type": "string",
            "maxLength": 512
          },
          "threadExternalId": {
            "type": "string",
            "maxLength": 512
          },
          "text": {
            "type": "string",
            "maxLength": 200000
          },
          "attachmentUrls": {
            "type": "array",
            "maxItems": 32,
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "receivedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "error": {
            "additionalProperties": false,
            "type": "object",
            "properties": {
              "version": {
                "const": 1,
                "type": "number"
              },
              "code": {
                "minLength": 1,
                "maxLength": 128,
                "type": "string"
              },
              "message": {
                "minLength": 1,
                "maxLength": 4096,
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              },
              "retryAfterMs": {
                "minimum": 0,
                "maximum": 86400000,
                "type": "integer"
              },
              "details": {
                "type": "object",
                "patternProperties": {
                  "^(.*)$": {}
                }
              }
            },
            "required": [
              "version",
              "code",
              "message",
              "retryable"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  },
  "security": [
    {
      "appToken": []
    }
  ],
  "paths": {
...
    "/hooks/webhook/{accountId}": {
      "post": {
        "security": [],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Lite-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^sha256=[a-fA-F0-9]{64}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboundEnvelope"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "accepted or duplicate delivery"
          },
          "401": {
            "description": "signature rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRun"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "accepted or replayed"
          }
        }
      }
    },
    "/v1/runs/{runId}": {
      "get": {
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "run"
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{runId}/events": {
      "get": {
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE event stream",
            "content": {
              "text/event-stream": {}
            }
          }
        }
      }
    },
    "/v1/runs/{runId}/attempts": {
      "get": {
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "durable execution attempts"
          }
        }
      }
    },
    "/v1/runs/{runId}/children": {
      "get": {
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "owned direct child runs"
          }
        }
      }
    },
    "/v1/runs/{runId}/cancel": {
      "post": {
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "cancelled or terminal run"
          }
        }
      }
    },
    "/v1/runs/{runId}/steer": {
      "post": {
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "instruction queued"
          }
        }
      }
    },
    "/v1/sessions/{sessionId}": {
      "get": {
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "session"
          }
        }
      }
    },
    "/v1/sessions/{sessionId}/messages": {
      "get": {
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "messages"
          }
        }
      }
    },
    "/v1/approvals/{approvalId}": {
      "post": {
        "parameters": [
          {
            "name": "approvalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "approval resolved"
          }
        }
      }
    },
    "/v1/runs/{runId}/artifacts": {
      "post": {
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "artifact published"
          }
        }
      }
    },
    "/v1/artifacts/{artifactId}": {
      "get": {
        "parameters": [
          {
            "name": "artifactId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "artifact metadata and base64 data"
          }
        }
      }
    },
    "/v1/agents": {
      "get": {
        "responses": {
          "200": {
            "description": "owned agent profiles"
          }
        }
      },
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgent"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "agent profile created"
          }
        }
      }
    },
    "/v1/agents/{agentId}": {
      "get": {
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "agent profile"
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces": {
      "get": {
        "responses": {
          "200": {
            "description": "owned workspaces"
          }
        }
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspace"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "managed workspace created"
          }
        }
      }
    },
    "/v1/workspaces/{workspaceId}": {
      "get": {
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "workspace"
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  },
...
}